+++ src/App.tsx
@@ -3,13 +3,12 @@ import './App.scss';
 import todosFromServer from './api/todos';
 import { TodoList } from './components/TodoList';
 import { Todo } from './types/Todo';
-import { TodoForm } from './components/TodoForm';
-import { getUserById } from './services/user';
 import { useState } from 'react';
+import usersFromServer from './api/users';
 
 const initialTodo: Todo[] = todosFromServer.map(todo => ({
   ...todo,
-  user: getUserById(todo.userId),
+  user: usersFromServer.find(user => user.id === todo.userId),
 }));
 
 function getNewTodoId(todos: Todo[]) {
@@ -19,6 +18,10 @@ function getNewTodoId(todos: Todo[]) {
 }
 
 export const App: React.FC = () => {
+  const [title, setTitle] = useState('');
+  const [hasTitleError, setHasTitleError] = useState(false);
+  const [userId, setUserId] = useState(0);
+  const [hasUserIdError, setHasUserIdError] = useState(false);
   const [todos, setTodos] = useState<Todo[]>(initialTodo);
 
   const addTodo = ({ id, ...data }: Todo) => {
@@ -30,11 +33,85 @@ export const App: React.FC = () => {
     setTodos(currentTodos => [...currentTodos, newTodo]);
   };
 
+  const handleTitleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setTitle(event.target.value);
+    setHasTitleError(false);
+  };
+
+  const handleUserIdChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
+    setUserId(+event.target.value);
+    setHasUserIdError(false);
+  };
+
+  const handleSubmit = (event: React.FormEvent) => {
+    event.preventDefault();
+
+    setHasTitleError(!title.trim());
+    setHasUserIdError(!userId);
+
+    if (!title || !userId) {
+      return;
+    }
+
+    addTodo({
+      id: getNewTodoId(todos),
+      title,
+      completed: false,
+      userId,
+      user: usersFromServer.find(user => user.id === userId),
+    });
+
+    setTitle('');
+    setUserId(0);
+    setHasTitleError(false);
+    setHasUserIdError(false);
+  };
+
   return (
     <div className="App">
       <h1>Add todo form</h1>
 
-      <TodoForm onSubmit={addTodo} />
+      <form onSubmit={handleSubmit} action="/api/todos" method="POST">
+        <div className="field">
+          <label htmlFor="post-title">Title: </label>
+          <input
+            id="post-title"
+            type="text"
+            data-cy="titleInput"
+            placeholder="Enter a title"
+            value={title}
+            onChange={handleTitleChange}
+          />
+          {hasTitleError && <span className="error">Please enter a title</span>}
+        </div>
+
+        <div className="field">
+          <label htmlFor="post-user">User: </label>
+          <select
+            id="post-user"
+            data-cy="userSelect"
+            value={userId}
+            onChange={handleUserIdChange}
+          >
+            <option value="0" disabled>
+              Choose a user
+            </option>
+            {usersFromServer.map(user => (
+              <option key={user.id} value={user.id}>
+                {user.name}
+              </option>
+            ))}
+          </select>
+
+          {hasUserIdError && (
+            <span className="error">Please choose a user</span>
+          )}
+        </div>
+
+        <button type="submit" data-cy="submitButton">
+          Add
+        </button>
+      </form>
 
       <TodoList todos={todos} />
     </div>
