--- src/components/TodoForm/TodoForm.tsx
@@ -1,91 +1,0 @@
-import { useState } from 'react';
-import usersFromServer from '../../api/users';
-import { Todo } from '../../types/Todo';
-import { getUserById } from '../../services/user';
-
-type Props = {
-  onSubmit: (todo: Todo) => void;
-};
-
-export const TodoForm: React.FC<Props> = ({ onSubmit }) => {
-  const [title, setTitle] = useState('');
-  const [hasTitleError, setHasTitleError] = useState(false);
-  const [userId, setUserId] = useState(0);
-  const [hasUserIdError, setHasUserIdError] = useState(false);
-
-  const handleTitleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setTitle(event.target.value);
-    setHasTitleError(false);
-  };
-
-  const handleUserIdChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
-    setUserId(+event.target.value);
-    setHasUserIdError(false);
-  };
-
-  const handleSubmit = (event: React.FormEvent) => {
-    event.preventDefault();
-
-    setHasTitleError(!title);
-    setHasUserIdError(!userId);
-
-    if (!title || !userId) {
-      return;
-    }
-
-    onSubmit({
-      id: 0,
-      title,
-      completed: false,
-      userId,
-      user: getUserById(userId),
-    });
-
-    setTitle('');
-    setUserId(0);
-    setHasTitleError(false);
-    setHasUserIdError(false);
-  };
-
-  return (
-    <form onSubmit={handleSubmit} action="/api/todos" method="POST">
-      <div className="field">
-        <label htmlFor="post-title">Title: </label>
-        <input
-          id="post-title"
-          type="text"
-          data-cy="titleInput"
-          placeholder="Enter a title"
-          value={title}
-          onChange={handleTitleChange}
-        />
-        {hasTitleError && <span className="error">Please enter a title</span>}
-      </div>
-
-      <div className="field">
-        <label htmlFor="post-user">User: </label>
-        <select
-          id="post-user"
-          data-cy="userSelect"
-          value={userId}
-          onChange={handleUserIdChange}
-        >
-          <option value="0" disabled>
-            Choose a user
-          </option>
-          {usersFromServer.map(user => (
-            <option key={user.id} value={user.id}>
-              {user.name}
-            </option>
-          ))}
-        </select>
-
-        {hasUserIdError && <span className="error">Please choose a user</span>}
-      </div>
-
-      <button type="submit" data-cy="submitButton">
-        Add
-      </button>
-    </form>
-  );
-};
