--- src/types/User.ts
@@ -1,6 +1,0 @@
-export type User = {
-  id: number;
-  name: string;
-  username: string;
-  email: string;
-};
